+++ packages/core/src/state-management/repository.ts
@@ -75,7 +75,12 @@ export class Repository {
   ) {
     this.loadingQ = new ExecutionQueue(concurrencyLimit, logger)
     this.executionQ = new ExecutionQueue(concurrencyLimit, logger)
-    this.inmemory = new StateCache(cacheLimit, (state$) => state$.complete())
+    this.inmemory = new StateCache(cacheLimit, (state$) => {
+      if (state$.subscriptionSet.size > 0) {
+        logger.debug(`Stream ${state$.id} evicted from cache while having subscriptions`)
+      }
+      state$.complete()
+    })
     this.updates$ = this.updates$.bind(this)
   }
 
@@ -307,6 +312,27 @@ export class Repository {
    */
   async listPinned(streamId?: StreamID): Promise<string[]> {
     return this.#deps.pinStore.ls(streamId)
+  }
+
+  /**
+   * Returns the StreamState of a random pinned stream from the state store
+   */
+  async randomPinnedStreamState(): Promise<StreamState | null> {
+    // First get a random streamID from the state store.
+    const res = await this.#deps.pinStore.stateStore.list(null, 1)
+    if (res.length == 0) {
+      return null
+    }
+    if (res.length > 1) {
+      // This should be impossible and indicates a programming error with how the state store
+      // list() call is enforcing the limit argument.
+      throw new Error(
+        `Expected a single streamID from the state store, but got ${res.length} streamIDs instead`
+      )
+    }
+
+    const [streamID] = res
+    return this.#deps.pinStore.stateStore.load(StreamID.fromString(streamID))
   }
 
   /**