+++ packages/blockchain-utils-linking/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.0.1-rc.0 (2022-04-27)
+
+**Note:** Version bump only for package @ceramicnetwork/blockchain-utils-linking
+
+
+
+
+
+# [2.0.0](/compare/@ceramicnetwork/blockchain-utils-linking@2.0.0-rc.3...@ceramicnetwork/blockchain-utils-linking@2.0.0) (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/blockchain-utils-linking
+
+
+
+
+
+# [2.0.0-rc.3](/compare/@ceramicnetwork/blockchain-utils-linking@2.0.0-rc.2...@ceramicnetwork/blockchain-utils-linking@2.0.0-rc.3) (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/blockchain-utils-linking
+
+
+
+
+
+# [2.0.0-rc.2](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/blockchain-utils-linking@2.0.0-rc.1...@ceramicnetwork/blockchain-utils-linking@2.0.0-rc.2) (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/blockchain-utils-linking
+
+
+
+
+
+# 2.0.0-rc.1 (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/blockchain-utils-linking
+
+
+
+
+
+# 2.0.0-rc.0 (2022-03-31)
+
+**Note:** Version bump only for package @ceramicnetwork/blockchain-utils-linking
+
+
+
+
+
+# [2.0.0-alpha.4](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/blockchain-utils-linking@2.0.0-alpha.3...@ceramicnetwork/blockchain-utils-linking@2.0.0-alpha.4) (2022-03-24)
+
+
+### Bug Fixes
+
+* Quick fix: OMG Forgot to update a snapshot file, it seems. ([7673020](https://github.com/ceramicnetwork/js-ceramic/commit/7673020a0e0a329efefcdcaf94341cb84fa16657))
+
+
+
+
+
 # [2.0.0-alpha.3](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/blockchain-utils-linking@2.0.0-alpha.2...@ceramicnetwork/blockchain-utils-linking@2.0.0-alpha.3) (2022-03-03)