+++ packages/key-did-resolver/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](/compare/key-did-resolver@2.0.0-rc.1...key-did-resolver@2.0.0) (2022-04-19)
+
+**Note:** Version bump only for package key-did-resolver
+
+
+
+
+
+# [2.0.0-rc.1](/compare/key-did-resolver@2.0.0-rc.0...key-did-resolver@2.0.0-rc.1) (2022-04-19)
+
+**Note:** Version bump only for package key-did-resolver
+
+
+
+
+
+# 2.0.0-rc.0 (2022-03-31)
+
+**Note:** Version bump only for package key-did-resolver
+
+
+
+
+
+# [2.0.0-alpha.4](https://github.com/ceramicnetwork/js-ceramic/compare/key-did-resolver@2.0.0-alpha.3...key-did-resolver@2.0.0-alpha.4) (2022-03-24)
+
+
+### Features
+
+* Do not depend on bigint-mod-arith ([#2074](https://github.com/ceramicnetwork/js-ceramic/issues/2074)) ([ad731cd](https://github.com/ceramicnetwork/js-ceramic/commit/ad731cd875ba4c1d9f398e4dc8be843ee4810862))
+* polyfill AbortController, so that Ceramic node works on Node.js v14 ([#2090](https://github.com/ceramicnetwork/js-ceramic/issues/2090)) ([fff3e8a](https://github.com/ceramicnetwork/js-ceramic/commit/fff3e8a18ef7d2ba86c80743f61f0487dae3e129))
+
+
+
+
+
 # [2.0.0-alpha.3](https://github.com/ceramicnetwork/js-ceramic/compare/key-did-resolver@2.0.0-alpha.2...key-did-resolver@2.0.0-alpha.3) (2022-03-03)