+++ packages/logger/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](/compare/@ceramicnetwork/logger@2.0.0-rc.1...@ceramicnetwork/logger@2.0.0) (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/logger
+
+
+
+
+
+# [2.0.0-rc.1](/compare/@ceramicnetwork/logger@2.0.0-rc.0...@ceramicnetwork/logger@2.0.0-rc.1) (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/logger
+
+
+
+
+
+# 2.0.0-rc.0 (2022-03-31)
+
+**Note:** Version bump only for package @ceramicnetwork/logger
+
+
+
+
+
 # [2.0.0-alpha.3](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/logger@2.0.0-alpha.2...@ceramicnetwork/logger@2.0.0-alpha.3) (2022-03-03)