+++ packages/stream-tile/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.0.1-rc.0 (2022-04-27)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile
+
+
+
+
+
+# [2.0.0](/compare/@ceramicnetwork/stream-tile@2.0.0-rc.4...@ceramicnetwork/stream-tile@2.0.0) (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile
+
+
+
+
+
+# [2.0.0-rc.4](/compare/@ceramicnetwork/stream-tile@2.0.0-rc.3...@ceramicnetwork/stream-tile@2.0.0-rc.4) (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile
+
+
+
+
+
+# [2.0.0-rc.3](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile@2.0.0-rc.2...@ceramicnetwork/stream-tile@2.0.0-rc.3) (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile
+
+
+
+
+
+# 2.0.0-rc.2 (2022-04-19)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile
+
+
+
+
+
+# [2.0.0-rc.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile@2.0.0-rc.0...@ceramicnetwork/stream-tile@2.0.0-rc.1) (2022-04-18)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile
+
+
+
+
+
+# 2.0.0-rc.0 (2022-03-31)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile
+
+
+
+
+
+# [2.0.0-alpha.4](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile@2.0.0-alpha.3...@ceramicnetwork/stream-tile@2.0.0-alpha.4) (2022-03-24)
+
+
+### Features
+
+* Add support for specifying resources by family ([#2070](https://github.com/ceramicnetwork/js-ceramic/issues/2070)) ([2a17f16](https://github.com/ceramicnetwork/js-ceramic/commit/2a17f16937702225d402ce13569e7ea0c3248990))
+
+
+
+
+
 # [2.0.0-alpha.3](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile@2.0.0-alpha.2...@ceramicnetwork/stream-tile@2.0.0-alpha.3) (2022-03-03)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-tile